Address Golden's Concerns
- Cast stone variables to save game compatible(I don't really need to do netsend since that is done at runtime).
- Increase stack size to match vanilla.
- Made ancient forges follow vanilla .automatically
- Remove 3rd hook in WormholeHacks since it appears to be redundant since I used it 25ish times in a server without fail.
- I needed to check the Void bag in WormholeHacks.
- Fix ring recipe.
- remove dependency from VanillaTweaks new solution is in said mod.

+++ Items/Consumable/LifeStone.cs
@@ -63,7 +63,7 @@ namespace GoldensMisc.Items.Consumable
 		
 		public override void NetSend(BinaryWriter writer)
 		{
-			writer.Write(uses);
+			writer.Write((byte)uses);
 		}
 		
 		public override void NetReceive(BinaryReader reader)
@@ -73,7 +73,7 @@ namespace GoldensMisc.Items.Consumable
 		
 		public override void SaveData(TagCompound tag)
 		{
-			tag["u"] = uses;
+			tag["u"] = (byte)uses;
 		}
 		
 		public override void LoadData(TagCompound tag)
